@@ -5,21 +5,23 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": "^7.2",
+        "php": "^7.3",
         "doctrine/dbal": "^2.10",
         "fideloper/proxy": "~4.0",
-        "laravel/framework": "~7.11.0",
+        "laravel/framework": "^8.0",
         "laravel/helpers": "^1.2",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
-        "twbs/bootstrap": "^3.3"
+        "laravel/ui": "^3.0",
+        "twbs/bootstrap": "^3.3",
+        "facade/ignition": "^2.3.6",
+        "guzzlehttp/guzzle": "^7.0.1"
     },
     "require-dev": {
         "filp/whoops": "~2.0",
         "fzaninotto/faker": "^1.9.1",
         "mockery/mockery": "^1.3",
-        "nunomaduro/collision": "^4.1",
-        "nunomaduro/larastan": "^0.5.8",
+        "nunomaduro/collision": "^5.0",
+        "nunomaduro/larastan": "^0.6.4",
         "phpunit/phpunit": "^9.0"
     },
     "autoload": {
