@@ -1,23 +1,30 @@
 <?php
 
-use Faker\Generator as Faker;
+namespace Database\Factories;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+use App\Catalog;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(App\Catalog::class, function (Faker $faker) {
-    return [
-        'name' => $faker->words(15, true),
-        'description' => $faker->paragraph,
-        'parent_id' => null,
-    ];
-});
+class CatalogFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Catalog::class;
 
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->words(15, true),
+            'description' => $this->faker->paragraph,
+            'parent_id' => null,
+        ];
+    }
+}
