@@ -1,25 +1,34 @@
 <?php
 
-use Faker\Generator as Faker;
+namespace Database\Factories;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+use App\Product;
+use Exception;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(App\Product::class, function (Faker $faker) {
-    return [
-        'name' => $faker->sentence,
-        'catalog_id' => random_int(1, 10),
-        'description' => $faker->paragraph,
-        'price' => $faker->randomFloat(2, 0.01, 9999999.99),
-        'image' => 'product_img.jpg'
-    ];
-});
+class ProductFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Product::class;
 
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     * @throws Exception
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->sentence,
+            'catalog_id' => random_int(1, 10),
+            'description' => $this->faker->paragraph,
+            'price' => $this->faker->randomFloat(2, 0.01, 9999999.99),
+            'image' => 'product_img.jpg'
+        ];
+    }
+}
